// src/app/containers/annuity/state/annuity.type.ts
export enum AnnuitiesActionTypes {
  SetCurrentAnnuity = '[Annuity] Set Current Annuity',
  ClearCurrentAnnuity = '[Annuity] Clear Current Annuity',
  InitializeCurrentAnnuity = '[Annuity] Initialize Current Annuity',

  Load = '[Annuity] Load',
  LoadSuccess = '[Annuity] Load Success',
  LoadFail = '[Annuity] Load Fail',

  UpdateAnnuity = '[Annuity] Update Annuity',
  UpdateAnnuitySuccess = '[Annuity] Update Annuity Success',
  UpdateAnnuityFail = '[Annuity] Update Annuity Fail',

  CreateAnnuity = '[Annuity] Create Annuity',
  CreateAnnuitySuccess = '[Annuity] Create Annuity Success',
  CreateAnnuityFail = '[Annuity] Create Annuity Fail',

  DeleteAnnuity = '[Annuity] Delete Annuity',
  DeleteAnnuitySuccess = '[Annuity] Delete Annuity Success',
  DeleteAnnuityFail = '[Annuity] Delete Annuity Fail'
}